refactor: colocate page styles with pages
pageExtensions config enables this possibility
